requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
PyMuPDF==1.23.8
Pillow==10.1.0
python-magic-bin==0.4.14
requests==2.32.4 